add some events
