+++ RustPlusApi/RustPlusApi.Fcm/Data/Events/ServerEventArg.cs
@@ -1,0 +1,9 @@
+﻿
+namespace RustPlusApi.Fcm.Data.Events
+{
+    public class ServerEventArg
+    {
+        public Guid Id { get; set; }
+        public string Name { get; set; } = null!;
+    }
+}
